CI:CD for Machine Learning using GitHub Actions/source/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asn1crypto==1.5.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==4.2.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cloudpickle==2.0.0
comm==0.1.3
contourpy==1.1.0
cryptography==38.0.4
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
gast==0.4.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.9.0
idna==3.4
imbalanced-learn==0.10.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.23.3
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.1.1
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.1
multidict==6.0.4
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
oscrypto==1.3.0
overrides==7.3.1
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.18.0
Pygments==2.15.1
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-tabnet==3.1.1
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
snowflake-connector-python==3.0.4
snowflake-ml-python==1.0.2
snowflake-snowpark-python==1.5.1
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0rc1
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0rc1
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
torch==1.10.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
uri-template==1.3.0
urllib3==1.26.16
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
widgetsnbextension==4.0.7
wrapt==1.15.0
xgboost==1.7.3
yarl==1.9.2
zipp==3.15.0
